Refactor validation utils for modularity and maintainability
Moved `check_unique_type` and `check_output_dim` to a new `validation_multi.py` module for better structure and separation of concerns. Updated imports accordingly to maintain backward compatibility.
